initial refactor: separate source and target datasets
- Works for ImageNet, COCO with Google Open Images
